+++ testing/Albatross.Testing/release-notes.md
@@ -1,4 +1,4 @@
 # 7.5.8
-* Add `DateOnlyExtensions` to support generation of random dates within a range.
+* Add `RandomExtensions` to support generation of random data.
 * Change the project to multi target .Net 6.0, 8.0